backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - Supabase
supabase==2.3.4
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client for external APIs
httpx>=0.24,<0.26

# Google Calendar API
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Validation and utilities
pydantic>=2.8.0
pydantic-settings>=2.3.0
email-validator==2.1.0

# Date/time handling
python-dateutil==2.8.2
pytz==2024.1
